Move startup diagnostics later in the startup chain
